+++ Source/GUI/Assimil8or/Editor/LoopPoints/LoopPointsView.cpp
@@ -21,7 +21,7 @@ void LoopPointsView::paint (juce::Graphics& g)
     {
         juce::dsp::AudioBlock<float> audioBlock { audioBuffer };
         juce::dsp::AudioBlock<float> loopSamples { audioBlock.getSubBlock (sampleOffset, numSamples) };
-        const auto samplesToDisplay {std::min<int>(numSamples, halfWidth)};
+        const auto samplesToDisplay { static_cast<int> (std::min<int64_t>(numSamples, halfWidth)) };
 
         g.setColour (juce::Colours::black);
         auto readPtr { loopSamples.getChannelPointer (0) };
@@ -30,12 +30,16 @@ void LoopPointsView::paint (juce::Graphics& g)
             // draw one line of sample going reverse from middle to left
             const auto xOffset { halfWidth - sampleCount };
             const auto sampleIndex { numSamples - sampleCount };
-            g.drawLine (xOffset, static_cast<int>(halfHeight - (readPtr [sampleIndex] * getHeight ())),
-                        xOffset + 1, static_cast<int>(halfHeight - (readPtr [sampleIndex + 1] * getHeight ())));
+            g.drawLine (static_cast<float> (xOffset),
+                        static_cast<float> (static_cast<int>(halfHeight - (readPtr [sampleIndex] * getHeight ()))),
+                        static_cast<float> (xOffset + 1),
+                        static_cast<float> (static_cast<int>(halfHeight - (readPtr [sampleIndex + 1] * getHeight ()))));
 
             // draw one line of sample start going from middle to right
-            g.drawLine (halfWidth + sampleCount, static_cast<int>(halfHeight - (readPtr [sampleCount] * getHeight ())),
-                        halfWidth + sampleCount + 1, static_cast<int>(halfHeight - (readPtr [sampleCount + 1] * getHeight ())));
+            g.drawLine (static_cast<float> (halfWidth + sampleCount),
+                        static_cast<float> (static_cast<int>(halfHeight - (readPtr [sampleCount] * getHeight ()))),
+                        static_cast<float> (halfWidth + sampleCount + 1),
+                        static_cast<float> (static_cast<int>(halfHeight - (readPtr [sampleCount + 1] * getHeight ()))));
         }
     }
